move the page already to presenter folder to start the implementation of clean arch
