+++ lib/app/repositories/presenter/repositories_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-import 'package:app_masterclass/components/header_page_widget.dart';
-import 'package:app_masterclass/app/repositories/components/repository_card.dart';
+import '../../../components/header_page_widget.dart';
+import 'components/repository_card.dart';
 
 class RepositoriesPage extends StatefulWidget {
   const RepositoriesPage({Key? key}) : super(key: key);
